perf: fix clippy warnings for better codegen
- Use next_back() instead of last() for DoubleEndedIterator
- Remove unnecessary u32 cast in benchmark.rs
- Zero clippy warnings now

+++ src/benchmark.rs
@@ -901,7 +901,7 @@ impl<'a> BenchmarkRunner<'a> {
             self.api_key,
             self.model,
             &full_prompt,
-            prompt.max_tokens as u32,
+            prompt.max_tokens,
         ).await?;
 
         Ok(response)
